+++ CHANGELOG.md
@@ -114,3 +114,8 @@
 * Updated to Ol8
 * Updated dev dependencies
 * Converted css to scss
+
+## v2.0.1
+* Added reframe and draw ROI functionality to improve the control of the region of interest
+* Removed scale parameter from the constructor and the settings modal
+* Removed some forgotten console.log